Issue #89 WIP
