+++ backend/src/main/java/de/njsm/versusvirus/backend/domain/Purchase.java
@@ -47,12 +47,6 @@ public class Purchase {
 
     private Instant executionTime;
 
-    private byte[] receipt;                   // picture of receipt
-
-    private String receiptMimeType;
-
-    private String receiptFileExtension;
-
     private BigDecimal cost;                  // cost of purchase in "Rappen"
 
     private boolean expensesPaid;             // 10.- per purchase by foundation (if Volunteer wantsCompensation)
@@ -73,16 +67,7 @@ public class Purchase {
     private List<Long> volunteerApplications;
 
     // telegram parameters
-    private String receiptFileId;
     private Long broadcastMessageId;
-
-    public String getReceiptMimeType() {
-        return receiptMimeType;
-    }
-
-    public void setReceiptMimeType(String receiptMimeType) {
-        this.receiptMimeType = receiptMimeType;
-    }
 
     public void writeToCsv(CSVPrinter csvPrinter, Optional<Customer> customer,
                            Optional<Volunteer> volunteer) throws IOException {
@@ -113,14 +98,6 @@ public class Purchase {
                 customer.map(Customer::getAddress).flatMap(Address::getZipCode).orElse(""),
                 customer.map(Customer::getAddress).map(Address::getCity).orElse("")
         );
-    }
-
-    public String getReceiptFileExtension() {
-        return receiptFileExtension;
-    }
-
-    public void setReceiptFileExtension(String receiptFileExtension) {
-        this.receiptFileExtension = receiptFileExtension;
     }
 
     public String getInternalComments() {
@@ -312,14 +289,6 @@ public class Purchase {
         this.privateComments = privateComments;
     }
 
-    public byte[] getReceipt() {
-        return receipt;
-    }
-
-    public void setReceipt(byte[] receipt) {
-        this.receipt = receipt;
-    }
-
     public Optional<BigDecimal> getCost() {
         return Optional.ofNullable(cost);
     }
@@ -334,14 +303,6 @@ public class Purchase {
 
     public void setExpensesPaid(boolean expensesPaid) {
         this.expensesPaid = expensesPaid;
-    }
-
-    public String getReceiptFileId() {
-        return receiptFileId;
-    }
-
-    public void setReceiptFileId(String receiptFileId) {
-        this.receiptFileId = receiptFileId;
     }
 
     public long getBroadcastMessageId() {
