+++ backend/src/main/java/de/njsm/versusvirus/backend/domain/PurchaseSupermarket.java
@@ -19,6 +19,14 @@ public class PurchaseSupermarket {
     @OneToMany(mappedBy = "purchaseSupermarket", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<OrderItem> purchaseList = new ArrayList<>();
 
+    private byte[] receipt;
+
+    private String receiptFileExtension;
+
+    private String receiptMimeType;
+
+    private String receiptFileId;
+
     public long getId() {
         return id;
     }
@@ -50,5 +58,37 @@ public class PurchaseSupermarket {
     public void addOrderItem(OrderItem purchaseItem) {
         purchaseList.add(purchaseItem);
         purchaseItem.setPurchaseSupermarket(this);
+    }
+
+    public byte[] getReceipt() {
+        return receipt;
+    }
+
+    public void setReceipt(byte[] receipt) {
+        this.receipt = receipt;
+    }
+
+    public String getReceiptFileExtension() {
+        return receiptFileExtension;
+    }
+
+    public void setReceiptFileExtension(String receiptFileExtension) {
+        this.receiptFileExtension = receiptFileExtension;
+    }
+
+    public String getReceiptMimeType() {
+        return receiptMimeType;
+    }
+
+    public void setReceiptMimeType(String receiptMimeType) {
+        this.receiptMimeType = receiptMimeType;
+    }
+
+    public String getReceiptFileId() {
+        return receiptFileId;
+    }
+
+    public void setReceiptFileId(String receiptFileId) {
+        this.receiptFileId = receiptFileId;
     }
 }
