+++ backend/src/main/java/de/njsm/versusvirus/backend/rest/api/admin/purchase/PurchaseController.java
@@ -78,7 +78,7 @@ public class PurchaseController {
 
     @GetMapping("/{id}/receipt")
     public ResponseEntity<byte[]> getReceipt(@PathVariable("id") UUID purchaseId) {
-        var image = purchaseService.getReceipt(purchaseId);
+        var image = purchaseService.getReceipts(purchaseId);
         return ResponseEntity.ok()
                 .header("Content-Disposition", "inline; filename=\"receipt." + image.getExtension() + "\"")
                 .contentType(org.springframework.http.MediaType.parseMediaType(image.getMimeType()))
